+++ gui_platform.hpp
@@ -17,6 +17,8 @@
 #include <commctrl.h>
 #include <commdlg.h>
 #include <shellapi.h>
+#include <uxtheme.h>
+#include <vssym32.h>
 
 #include <algorithm>
 #include <atomic>